backend/requirements.txt
Django
djangorestframework
pandas
reportlab
django-cors-headers
openpyxl
djangorestframework-simplejwt
